+++ android/src/main/java/com/odehbros/flutter_file_downloader/MethodCallHandlerImpl.java
@@ -150,10 +150,12 @@ public class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
         String lastURL = helper.call.argument("url");
         String lastName = helper.call.argument("name");
         String lastDestination = helper.call.argument("download_destination");
+        String notifications = helper.call.argument("notifications");
 
         new DownloadTaskBuilder(activity)
                 .setUrl(lastURL)
                 .setName(lastName)
+                .setShowNotifications(notifications)
                 .setDownloadDestination(lastDestination)
                 .setCallbacks(new DownloadCallbacks() {
                     @Override
@@ -233,7 +235,7 @@ public class MethodCallHandlerImpl implements MethodChannel.MethodCallHandler {
                     }
                 })
                 .build()
-                .startDownloading();
+                .startDownloading(context);
     }
 
     public void removeTask(final long id) {