+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 1.1.3-dev.3
+## 1.1.3-dev.4
 
-- Fixed build exception
-- Fixed issue when download is complete, the native code does not return the file path back to flutter
+- Added the ability to manage download notifications, you can now enable all notifications, disabled all or use the default (only progress notifications)
+- Cleaned some code and enhancements